requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0 